# [perf] Add Equals implementation to the TransformedBounds