## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->
Implements equality function for `TransformedBounds`.

## What is the current behavior?
<!--- If the PR is a fix, describe the current incorrect behavior, otherwise delete this section. -->
Due to lack of equality function the runtime is using reflection to do it. Which is a on the slow side as it has to allocate a bit more as well. The benchmark is coming from setting `TransformedBounds` on a `Visual`, so it is not 100% accurate, but it gives some idea about the effect of this PR.


`Transformed bounds baseline`

|            Method |     Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------ |---------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
| TransformedBounds | 1.409 ms | 0.0049 ms | 0.0046 ms |   1080.0781 |           - |           - |           830.67 KB |

`Transformed bounds - equals implemented`

|            Method |     Mean |    Error |   StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------ |---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
| TransformedBounds | 406.4 us | 1.585 us | 1.482 us |    735.8398 |           - |           - |            565.3 KB |

## What is the updated/expected behavior with this PR?
<!--- Describe how to test the PR. -->
When checking for equality of `TransformedBounds` usage of reflection is avoided.

## How was the solution implemented (if it's not obvious)?
<!--- Include any information that might be of use to a reviewer here. -->


## Checklist

- [ ] Added unit tests (if possible)?
- [ ] Added XML documentation to any related classes?
- [ ] Consider submitting a PR to https://github.com/AvaloniaUI/Avaloniaui.net with user documentation

## Breaking changes
<!--- List any breaking changes here. When the PR is merged please add an entry to https://github.com/AvaloniaUI/Avalonia/wiki/Breaking-Changes -->


## Fixed issues
<!--- If the pull request fixes issue(s) list them like this: 
Fixes #123
Fixes #456
-->